# doc/api.yaml
openapi: 3.0.0
paths:
...
  /artist:
    post:
      operationId: ArtistController_create
      summary: Create artist
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateArtistDto'
      responses:
        '201':
          description: ''
        '400':
          description: ''
        '404':
          description: ''
      tags:
        - Artist
    get:
      operationId: ArtistController_findAll
      summary: Get all artist
      parameters: []
      responses:
        '200':
          description: ''
      tags:
        - Artist
  /artist/{id}:
    get:
      operationId: ArtistController_findOne
      summary: Get artist by id
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
        '400':
          description: ''
        '404':
          description: ''
      tags:
        - Artist
    put:
      operationId: ArtistController_update
      summary: Update artist
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateArtistDto'
      responses:
        '200':
          description: ''
        '400':
          description: ''
        '404':
          description: ''
      tags:
        - Artist
    delete:
      operationId: ArtistController_remove
      summary: Delete artist
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: ''
        '400':
          description: ''
        '404':
          description: ''
      tags:
        - Artist
...
  /favs/artist/{id}:
    post:
      operationId: FavoritesController_addArtist
      summary: Add artist to favorites
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '201':
          description: ''
        '404':
          description: ''
        '422':
          description: ''
      tags:
        - Favorite
    delete:
      operationId: FavoritesController_removeArtist
      summary: Delete artist from favorites
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: ''
        '400':
          description: ''
        '404':
          description: ''
      tags:
        - Favorite
info:
  title: "\tHome Library Service"
  description: 'Assignment: REST Service'
  version: '1.0'
  contact: {}
...
servers: []
components:
  schemas:
...
    CreateArtistDto:
      type: object
      properties:
        name:
          type: string
        grammy:
          type: boolean
      required:
        - name
        - grammy
    UpdateArtistDto:
      type: object
      properties:
        name:
          type: string
        grammy:
          type: boolean
...